feat!: remove MongoDB implementations (moved to event-mongodb)
All MongoDB-specific implementations have been moved to the dedicated
github.com/rbaliyan/event-mongodb module. This removes the
go.mongodb.org/mongo-driver/v2 dependency from the core library.

Deleted packages:
- checkpoint/mongodb.go       → event-mongodb/checkpoint
- distributed/mongodb.go      → event-mongodb/distributed
- distributed/mongodb_worker_store.go → event-mongodb/distributed
- idempotency/mongodb.go      → event-mongodb/idempotency
- monitor/mongodb.go          → event-mongodb/monitor
- outbox/mongodb.go           → event-mongodb/outbox
- outbox/relay_mongo.go       → event-mongodb/outbox
- outbox/relay_mongo_changestream.go → event-mongodb/outbox
- payload/bson.go             → event-mongodb/payload
- schema/mongodb.go           → event-mongodb/schema
- store/mongodb.go            → event-mongodb/store
- transaction/mongodb.go      → event-mongodb/transaction
- transport/codec/bson.go     → event-mongodb/codec

Added:
- distributed/worker_store_cursor.go: extracted workerCursor helpers
  (encodeWorkerCursor/decodeWorkerCursor) from the deleted mongodb_worker_store.go;
  these are pure JSON+base64 helpers still needed by memory_worker_store.go

Updated tests:
- outbox/outbox_test.go: removed Transaction/TransactionWithOptions tests
  (those helpers moved to event-mongodb/outbox)
- payload/payload_test.go: removed TestBSON, BenchmarkBSON, and BSON codec
  entries (codec moved to event-mongodb/payload)

go.mod: remove go.mongodb.org/mongo-driver/v2 and its transitive deps

BREAKING CHANGE: MongoDB implementations are no longer in this module.
Migrate to github.com/rbaliyan/event-mongodb sub-packages:

  event/v3/distributed.NewMongoStateManager  → event-mongodb/distributed
  event/v3/idempotency.NewMongoStore         → event-mongodb/idempotency
  event/v3/monitor.NewMongoStore             → event-mongodb/monitor
  event/v3/outbox.NewMongoStore              → event-mongodb/outbox
  event/v3/schema.NewMongoProvider           → event-mongodb/schema
  event/v3/checkpoint.NewMongoStore          → event-mongodb/checkpoint
  event/v3/transport/codec.BSON              → event-mongodb/codec
  event/v3/payload.BSON                      → event-mongodb/payload
  event/v3/store.NewMongoStore[T]            → event-mongodb/store
  event/v3/transaction.NewMongoManager       → event-mongodb/transaction
