+++ outbox/outbox_test.go
@@ -715,70 +715,3 @@ func TestPostgresStore_Store_WrongSessionType(t *testing.T) {
 	}
 }
 
-func TestMongoTransaction_PiggyBack(t *testing.T) {
-	// Set a context.Context session (simulating Mongo session context).
-	// client is nil — if piggy-back works, StartSession is never called.
-	ctx := event.WithOutboxTx(context.Background(), context.Background())
-
-	called := false
-	err := Transaction(ctx, nil, func(innerCtx context.Context) error {
-		called = true
-		return nil
-	})
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if !called {
-		t.Error("fn should have been called")
-	}
-}
-
-func TestMongoTransaction_PiggyBack_PropagatesError(t *testing.T) {
-	ctx := event.WithOutboxTx(context.Background(), context.Background())
-
-	testErr := errors.New("mongo business logic failed")
-	err := Transaction(ctx, nil, func(_ context.Context) error {
-		return testErr
-	})
-	if !errors.Is(err, testErr) {
-		t.Fatalf("expected %v, got %v", testErr, err)
-	}
-}
-
-func TestMongoTransactionWithOptions_PiggyBack(t *testing.T) {
-	ctx := event.WithOutboxTx(context.Background(), context.Background())
-
-	called := false
-	err := TransactionWithOptions(ctx, nil, nil, func(_ context.Context) error {
-		called = true
-		return nil
-	})
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if !called {
-		t.Error("fn should have been called")
-	}
-}
-
-func TestMongoTransaction_NoPiggyBackOnSqlTx(t *testing.T) {
-	// Set a *sql.Tx session — Mongo Transaction should NOT piggy-back.
-	ctx := event.WithOutboxTx(context.Background(), &sql.Tx{})
-
-	// client is nil, so StartSession will panic — proves piggy-back was skipped.
-	panicked := false
-	func() {
-		defer func() {
-			if r := recover(); r != nil {
-				panicked = true
-			}
-		}()
-		Transaction(ctx, nil, func(_ context.Context) error {
-			t.Fatal("fn should not have been called")
-			return nil
-		})
-	}()
-	if !panicked {
-		t.Fatal("expected panic when client is nil and piggy-back is skipped")
-	}
-}
